replayer/pyproject.toml
[project]
name = "replayer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "confluent-kafka>=2.12.2",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
]
